feat: reorganized baton router and handlers for improved structure and clarity
