+++ handlers/mparticle-api/test/baton.spec.ts
@@ -1,5 +1,8 @@
 import { faker } from '@faker-js/faker';
-import type { GUID, InitiationReference } from '../src/routers/baton';
+import type {
+	GUID,
+	InitiationReference,
+} from '../src/routers/baton/types-and-schemas';
 import type { AppConfig } from '../src/utils/config';
 import { invokeBatonRerHandler } from '../src/utils/invoke-baton-rer-handler';
